+++ src/components/time/time.pw.tsx
@@ -434,7 +434,7 @@ test.describe("Time component", () => {
     });
   });
 
-  test.describe("Accessibility tests ", () => {
+  test.describe("Accessibility tests", () => {
     test("should pass for default implementation", async ({ mount, page }) => {
       await mount(
         <TimeComponent